Moving land and take off to the airport class
