+++ lib/plane.rb
@@ -1,17 +1,6 @@
-require_relative 'weather'
+
+require_relative 'airport'
 require 'pry'
 
 class Plane
-  def initialize
-  end
-
-  def land(weather)
-    fail "The weather is stormy" if weather == "stormy"
-    "The plane has landed"
-  end
-
-  def take_off(weather)
-    fail "The weather is stormy" if weather == "stormy"
-    "The plane has taken off"
-  end
 end
